requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_07078715-3ab7-4562-8d3d-d56b0eaa0f7dp504n_ny/croots/recipe/bottleneck_1657175566567/work
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
faiss==1.7.4
filelock==3.13.1
Flask==3.0.2
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
markdown2==2.4.12
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numexpr @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_45yefq0kt6/croot/numexpr_1696515289183/work
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cfbbnu3tzv/croot/numpy_and_numpy_base_1704311724719/work/dist/numpy-1.26.3-cp310-cp310-macosx_11_0_arm64.whl#sha256=ccbf655215ec644fb09de935330d2602e8101b485f43bc98fefdd9ada37b9193
openai==1.11.0
packaging==23.2
pandas @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_82r7p4atwy/croot/pandas_1702318000240/work/dist/pandas-2.1.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=c8b7db467b7e229a08d8c41ceef1d279ae19ed4c6deb6d4dc494a3a64d00cddf
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.1
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6btwyyj8a1/croot/pytz_1695131592184/work
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.5
sympy==1.12
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.2.0
Werkzeug==3.0.1
